+++ DaisyAPP/app/src/main/java/PlantAdapter.kt
@@ -1,5 +1,3 @@
-import android.content.Context
-import android.graphics.BitmapFactory
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
@@ -8,8 +6,7 @@ import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.daisyapp.R
-import java.io.File
+import com.GoToExpress.daisyapp.R
 
 class PlantAdapter(private var plants: List<Plant>,private val listener: OnItemClickListener) : RecyclerView.Adapter<PlantAdapter.PlantViewHolder>() {
