+++ DaisyAPP/app/src/main/java/PlantSearchAdapter.kt
@@ -3,7 +3,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.example.daisyapp.R
+import com.GoToExpress.daisyapp.R
 
 class PlantSearchAdapter(
     private val onPlantSelected: (Plant) -> Unit
