+++ DaisyAPP/app/src/main/res/drawable/book_86162.xml
@@ -1,0 +1,5 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="48dp" android:viewportHeight="594.11" android:viewportWidth="594.11" android:width="48dp">
+      
+    <path android:fillColor="#FF000000" android:pathData="M543.93,425.96c0,5.94 -4.98,10.77 -11.18,10.77H381.25c-53.92,0 -73.49,24.33 -73.49,24.33l-2.49,3.34l-1.16,1.62c-0.01,0 -0.01,0 -0.01,0c-5.14,6.98 -12.7,0.05 -12.74,0l-3.49,-4.34c-2.59,-3.11 -22.4,-24.95 -70.32,-24.95H61.34c-6.17,0 -11.15,-4.85 -11.15,-10.79v-322.29c0,-5.96 4.99,-10.76 11.15,-10.76h121.27c90.21,0 98.75,51.24 99.34,66.11v224.15c0,16.29 7.76,19.33 12.14,19.78h6.52c4.46,-0.44 11.56,-3.1 11.56,-15.79V156.13h-0.08c0.69,-18.64 7.94,-63.24 62.2,-63.24h158.46c6.2,0 11.18,4.81 11.18,10.76V425.96L543.93,425.96zM584.07,156.14h-0.53c-10.18,0.06 -11.76,-8.8 -11.93,-12.9V78.96c0,-6.85 -5.57,-12.41 -12.41,-12.41H375.43c-46.12,0 -63.31,17.08 -67.58,22.46l-1.17,1.62l-0.01,0.02l-3.64,4.99l0,0c-4.88,6.71 -9.13,3.84 -11.19,1.65l-6.96,-8.68l-3.43,-4.14c-5.78,-6.07 -20.85,-17.93 -50.72,-17.93H34.94c-6.86,0 -12.42,5.56 -12.42,12.41v63.47c0,11.1 -6.03,13.34 -9.73,13.71h-2.76C7.01,156.31 0,158.07 0,170.97v302.42c0,7.43 6.02,13.45 13.44,13.45h178.68c57.01,0 81.85,21.5 89.3,29.73l4.25,5.27h0.01c0,0 11.4,12.88 23.41,0l0,0l2.51,-2.93v0.04l0.02,-0.04l3.18,-3.72c0.09,-0.11 0.86,-0.96 1.96,-2.11c9.52,-9.25 33.29,-26.24 79.77,-26.24h184.15c7.42,0 13.42,-6.01 13.42,-13.45V168.04C594.12,157.12 586.19,156.2 584.07,156.14L584.07,156.14z"/>
+    
+</vector>
