+++ DaisyAPP/app/src/main/res/layout/activity_auth.xml
@@ -9,7 +9,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="com.example.daisyapp.SingInFragment"
+        android:name="com.GoToExpress.daisyapp.SingInFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toBottomOf="parent"
