+++ DaisyAPP/app/src/main/res/layout/fragment_about.xml
@@ -171,7 +171,6 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="12dp"
-                app:cardBackgroundColor="#F5FFF8"
                 app:cardCornerRadius="12dp"
                 app:cardElevation="0dp">
